A first attempt to move all ESP-specific code to esp/esp_local Massive reorganization of esp.esp_local.views (formerly esp.web.views), specifically .program() and .myesp().  Split into separate files.
WARNING: THIS CODE DOES NOT WORK AS-IS!!  I've split things so violently 
that, at the very least, urls.py is broken and a bunch of files need 
their "import" statements to be updated.

Also, permissions handling in esp.esp_local.programs.* hasn't been 
reinstated yet; it was pulled because it was ugly.


git-svn-id: http://esp.mit.edu/code/esp-project/trunk@335 9f9a36c0-4815-0410-8ed4-a58642b8ae3a