Fix the really nasty navBar bug(s) that caused navBar permissions to be badly borked and allowed much of the navBar table to get deleted.
Moved settings.py to settings_files/aseering_devserver.  I will check in 
a parallel copy for the main server.  Please keep your settings.py in 
this subdirectory, and symlink it to /settings.py (and don't check the 
symlink into svn!); the different dev servers are starting to require 
different settings.py files, so we have to split them somehow.


git-svn-id: http://esp.mit.edu/code/esp-project/trunk@538 9f9a36c0-4815-0410-8ed4-a58642b8ae3a