Updated username validation.
We now use a whitelist to only allow certain characters, instead of a blacklist to disallow characters.

We also allow a much more restricted set of characters.  If it's too restrictive, feel free to add characters to the "good_chars" set in "clean_username", though if you do, be sure to update the form text and function spec accordingly.


git-svn-id: http://esp.mit.edu/code/esp-project/trunk@1047 9f9a36c0-4815-0410-8ed4-a58642b8ae3a