Some more optimizations and added some caching to the dashboard page.  Reduced its load time by another order of magnitude, when fully cached.
Right now, almost all of the CPU time on this page comes from summing the sets of students in each timeslot, and almost all of the DB time comes from each line in Program Vitals (each is now a .count() query).  The only way to make those sections faster would be to cache their results in memory.


git-svn-id: http://esp.mit.edu/code/esp-project/trunk@1258 9f9a36c0-4815-0410-8ed4-a58642b8ae3a