Fixing QSD editing to save a new row, instead of updating, when using the AJAX editor. This should cause *logged-out* users to see accurate last updated dates and times.
Unfortunately, the deal for logged-in users is... more special.

Because users might have special ways to display stuff, our caching system includes the user in the cache key. From templatetags/render_qsd.py:
def cache_key(qsd, user=None):
    # IF you change this, update qsd/models.py's QSDManager class
    # Otherwise, the wrong cache path will be invalidated
    # Also, make sure the qsd/models.py's get_file_id method
    # is also updated. Otherwise, other things might break.
    if user:
        return md5.new('%s-%s-%s' % (qsd.path.uri, qsd.name, user.id)).hexdigest()
    else:
        return md5.new('%s-%s' % (qsd.path.uri, qsd.name)).hexdigest()

Note that the cache key for logged-in users is uri-name-id (we won't even talk about the comment above --- getting *consistent* keys today was hard, too...).

Unfortunately, this makes invalidating the cache hard. While the primary/anon cache key is invalidated in the save method of QSDs, invalidating all user caches would, because of the use of an MD5, requiring looping over every registered user (I think), calculating the cache key, and deleting it. This would be slow. I also considered not caching for logged in users, but I don't know how to tell the file cache not to cache something from the cache_key function, and besides, most users are logged-in (I suspect), so caching only anon users would hardly be worthwhile.

Oh, and what are we caching? This type of thing:
@cache_inclusion_tag(register,'inclusion/qsd/render_qsd.html', cache_key_func=cache_key, cache_obj=render_qsd_cache, cache_time=86400)
def render_qsd(qsd, user=None):
    edit_bits = False
    if user:
        edit_bits = UserBit.UserHasPerms(user, qsd.path, DataTree.get_by_uri('V/Administer/Edit'))
    return {'qsdrec': qsd, 'edit_bits': edit_bits}

Do we need to cache this? How computationally expensive is it? Can we shorten the cache lifetime? Can we do finer-grained caching, and cache the UserHasPerms call instead of the whole template tag?

Any suggestions would be appreciated.

(This also doesn't even mention file_db. Is a homegrown, pickle-based file DB actually faster than PostgreSQL for data access (which seems to be about all that file_db is used for --- it doesn't actually seem to be a computation cache...)?)



git-svn-id: http://esp.mit.edu/code/esp-project/branches/django-HEAD-transition@1715 9f9a36c0-4815-0410-8ed4-a58642b8ae3a