As discussed with Adam, change TeacherBio urls to teachers/USERNAME/bio.html. The old urls of teachers/LASTNAME/FIRSTNAME/USERNUM/bio.html still work.
Computing userNum is rather expensive, and this is done for every teacher on
the Splash catalog page. Also, the urls are not persistent if someone changes
their name. (In which case I don't think the cache knows to invalidate itself
either.)

With this change, the Splash catalog requires 530 fewer database queries
and takes 19 fewer seconds to generate on no cache. (From 8222q,89s to
7692q,70s on my machine.)



git-svn-id: http://esp.mit.edu/code/esp-project/branches/django-HEAD-transition@1723 9f9a36c0-4815-0410-8ed4-a58642b8ae3a