Don't set base_fields in constructor.
It makes weird things happen, should we ever subclass these and override fields.
(Actually, newforms is hopelessly broken and weird things happen anyway, but this
particular weirdness is the weirdest and can be avoided[*].)

[*] Long story short, newforms has lots of bits where it's very easy for parent field
definitions to override child definitions as opposed to the reverse. It's also really
broken in other ways, but never mind that.