Add SHA-1-based LaTeX implementation.
The file name of each image is now a deterministic function of its content and
attributes, namely a SHA-1 hash of the content, style, and dpi. This means that
LaTeX can be queried without any database queries, and nowhere does it store an
actual path, so things can't break that way.