Make the world happy.
Happiness is good. With frogs and exploding pigeons and lots of shiiiiiny things. Shiiiiinnnnyyyyy. Ooh, look! It's a flying hippo! It is checkerboard. I bet you could play chess on it. Oh man, I'm losing. That's not so good.

Oh, anyway, this makes verb_base and qsc_base assigned properly.