Don't pick a random filename for the histograms; cache them. They take too long to generate.
We currently put them in random places anyway without cleaning up, so this
isn't any worse, except that we don't need to regenerate them all the time.
Ideally we'd offload this to Google Charts or something... or AT LEAST have
some generic SHA-1-based caching system. I'm working on that...

The filename is a hash of the pickled context dictionary.