It's not runnin'! It's passed on!
This function is no more! It has ceased to be! It's expired and gone to meet
its maker! It's a stiff! Bereft of use, it rests in peace! If you hadn't nailed
it to the module it'd be pushing up the daisies! Its running processes are now
history! It's off the repo! It's kicked the bit-bucket, it's shuffled off its
cpu core, run down the curtain and joined the bleedin' choir invisible!! THIS
IS AN EX-FUNCTION!!

*ahem*, removed save_instance.