Don't try to select every student when no students are specified
So, it turns out Django has this really obnoxious special-casing of the empty
query ( Q() ) which presumably tries to make your life easier, but actually
obscures potentially critical bugs. Imagine if this were a bulk-delete-user
page instead of a bulk-enroll-students page. As it is, we were only trying to
enroll all of ESP into AP Bio.

Moral of the story: Magic is bad. See also, Python descriptor protocol, Python
type system, and pretty much the rest of Python.