Make ArgCache's conform to new super semantics
Because multiple-inheritance in Python is horribly broken and
super()-driven cooperative calling is the stupidest thing ever. And it's
apparently causing problems in Python 2.6. It's currently not necessary,
as we don't use multiple inheritance, but I'm not surprised if we start
doing so at some point... if nothing else because it's a nice way to
separate ArgCache up.