Replace instances of media_url with MEDIA_URL
This particular change, unfortunately, will not merge well, due to
template breakage.  Mike Price, after you get around to merging the
other branches, *get rid of main.html*! The semantics of changes to the
main branch are "these changes should go to every branch" which means
having it use a different template name is harmful to the project.

Although main no longer needs it, DO NOT REMOVE
esp.context_processors.media_url. Because of the broken template
situation, we should not do this until everyone is merged past this
point and then this change is manually completed on every branch.

We should come up with a mechanism for these sorts of changes. The
template deletion means that this will be a common occurence.