Changes to k12school behavior in profile
Now the k12school field is no longer required.  However, if
the student leaves the field blank or enters and unrecognized string,
that string will be copied to the school field (which is non-required
and hidden in the template) and 'Other' will be saved as their k12school.

Also contains a fix to a profile error for educators and parents.