New approach to resource requests at teacher's class registration ...
A ResourceRequestFormSet has been added to the makeaclass page.
This formset allows a list of requests to be built up for each class.
Currently the formset is static, so you may not want to use this code
right away.  Forthcoming commits will use the Ajax tools to allow
adding/removing forms for different resource types.

Also added initial resource types and updated ProgramHappenTest.

There is a small database schema change for the resources models.