More complete resource requests for teacher reg form
Ajax views are provided for teachers to add/remove requests
for existing resource types and new resource types.  This is not
as simple or as complicated as it could be.

Also includes some improvements to the Ajax tools, and removes
mentions of the "Teacher Availability" resource type because
it is no longer used (teacher availability is stored in the
UserAvailability model).