Fix "Uninitialized cache." on server start.
It was kind of a circular import; an app in INSTALLED_APPS was causing
the rest of the apps to be imported before it had registered its caches.

The long story, as far as I can tell, was:
1. Django starts importing everything in INSTALLED_APPS.
2. The app "esp.program.modules" starts importing the program modules.
3. The onsite check-in module triggers a call to AppCache.get_models().
     (get_field_by_name() does it to find reverse relationships.)
4. AppCache.get_models() continues down the INSTALLED_APPS list.
5. The app "esp.cache_loader" checks that everything was imported.
     (esp.program.modules is "imported"; it just isn't done running.)
6. cache_loader locks the caches, thinking No Cache was Left Behind.
7. AppCache.get_models() finishes.
8. The program modules resume loading, now with the caches locked.
9. The ajax scheduling module tries to register some caches.

Mike Price's recent commit 8f1a3f69ab4e9b6b6b3f40f8e850225dff3fae43
(sorting the program modules) has pretty much already fixed this.
It turns out that the relevant module was onsitecheckinmodule.py, and
that the only program modules using @cache_function have names starting
with "a" and "e".