memcached may report concurrency errors if get_many/set_many invocations happen too close to each other.  Wrap these in a procedure that will retry up to (arbitrarily) 8 times before failing outright.
Note for the future:  Almost all of our read-based uses of the cache are safe if they declare a cache miss when there wasn't one.  (That's what python-memcached did.)  We may want to make this be the default behavior after a few retries, so that we don't error at all.  (Though we should log these cases in case we hit a memcached bug and start having no cache misses.)