Daemon to allow interoperating with the website
This script aims to provide a socket interface to the Django auth DB so that
other systems can connect to it and ask for people to be authenticated.
It's currently basically a copy of cmdline_auth.py, which provides similar
functionality except not as a persistent service.

This script has two main additional changes:
* Not optimized for Mediawiki
* Some additional queries (FINGER, CHECK USERBIT)