AJAX Scheduling: Refresh Directory each operation.
That is, re-filter the list of (unscheduled) sections each time a class
gets scheduled or unscheduled. It incurs a bit of overhead, but I think
it's worth it, to get an interface that obeys intuitive expectations.

I don't think this makes things appreciably slower, even for Splash.

I added a new jQuery event, in order to avoid race conditions:
    "block_section_assignment_success"
Sometimes Directory.filter() seemed to leave data.section undefined.
I'm not really sure why that was.