Fix more sadface errors.
[PART ONE]

Once again, the problem was with control escaping from
/esp/esp/program/modules/handlers/__init__.py before it finished
importing all the handlers.

It began in admincore.py at:
     43: from esp.datatree.forms import AjaxTreeField

In the definition of AjaxTreeField, we call DataTree.root().
As part of a member function definition, it gets evaluated when the
definition is interpreted (rather than on each call):
     48: def __init__(self, root=DataTree.root(), ...

Finally, in DataTree.root() we call DataTree.objects.get(...):
    512: cls.ROOT_NODE = cls.objects.get(name = cls.ROOT_NAME,
    513:                                 parent__isnull = True)

Django does some magic to set up *.objects. In particular it wants to
know about all the models that exist. So this makes Django interrupt our
loading of program modules to load the rest of the apps.

The rest is as before; cache_loader looks around, sees that all apps
have been loaded, and declares the caches finalized. Then program
modules resume loading and attempt to set up their caches.

Now if this were all, one could simply make the default None and check
for a value of None when __init__ gets executed. However...

[PART TWO]

In admincore.py, the definition of NewUserBitForm includes the line:
     54: verb = AjaxTreeField(label='Activity', root=GetNode(...

So GetNode has to be called when this definition is interpreted. Of
course, GetNode eventually calls DataTree.objects.get(...), so we run
into the same problem again.

Really, we'd be running into this every time we tried to use an
AjaxTreeField. So to sidestep the issue entirely, AjaxTreeField is now
defined using a URI. Only when we render the field do we actually
convert this into a node.