Allow default values for form fields to be specified in a Tag
Previously, we had used hard-coded initial values for states in the user profile forms.
In order to allow sharing the same code, this needs to be customizable at runtime.
So I added a Form mixin class, FormWithTagInitialValues.  If you make a Form class
inherit from FormWith
TagInitialValues, its constructor will accept a tag_map argument.
This argument should be a dictionary, where the keys are the names of fields to provide
initial values for and the values are the names of the Tags that those values should
come from.