Use Ajax field for K12School in student profile form
This is a variation on the behavior employed by Chicago.  Essentially it's an AjaxForeignKey field, but if there is no match (i.e. your school is not in the K12School database) you can still submit the form.

If you click on a school entry, that school is saved in the k12school field of the StudentInfo.  If you don't, the text you entered is saved in the school field of StudentInfo.