Fix graduation year handling.
The main branch stored the graduation year as an integer with
0 for N/A and 1 for 'graduate student'.  This was a bit confusing,
so I tried to simplify it by making the graduation_year a CharField
(you can enter whatever you want, but this probably won't be abused)
and adding an 'is_graduate_student' field.  The template for the profile
form has been changed to reflect this, but you can override it to specify
whatever directions you want teachers to follow.