Add PostgreSQL 9 support.
PostgreSQL 9 changes the PL/pgSQL language specification slightly; some function argument names had to be changed to make some expressions be unambiguous.

This also adds the 'datatree' model to the set of models managed by South / data migrations.  It adds a migration that will load these changes into the database.  The migration is entirely custom SQL; if it refuses to run (which will happen if your database permissions are wrong, in ways that don't affect anything else), you can just fake the migration and run the SQL yourself with "sudo -u postgres psql" or similar.