Disable CSRF protection on CyberSource postback.
It's not clear to me whether this is more secure than finagling some way
for the postback to include the CSRF token. Perhaps it is, since this
way we don't have to give our token to CyberSource's website--though if
we couldn't trust them we'd have bigger problems.

What *is* clear is that this is much easier.