Added initial statistics form and view; not pretty yet
The initial queries look up info on:
- demographics
- timeslots
- how students heard about ESP/Splash/etc.
- repeat participants
- schools
- zip codes
- timing of registration
(see http://wiki.learningu.org/Top_Statistics_Queries)

This is intended to be made into a prettier Ajax form, with some fields
dynamically hidden based on the query that is selected (and the checkboxes).
Additional queries can be added by placing an entry in StatisticsQueryForm
and the corresponding function in esp/program/statistics.py.