Fix up school/k12school integration on profile form.
Brief explanation: the k12school field is an AjaxForeignKey field
that is displayed, and the school field is hidden.  If there is
some error validating the k12school (such as if the user did not
click one of the dropdown entries, or there were no matches in
the k12school table for what they types), then the contents of
that text box will be saved in the school field.  When populating
initial values, the k12school will be used if available, with
school as a backup.