Fix get_user-related cache dumping (sad_face).
All cache functions need to be loaded before cache_loader runs.
It'll do its best to scan through once more for cache functions,
but it basically only looks in models.py files, since that's where
Django looks.

So we import get_user at the bottom of the nearest models file.
I realize that this makes the comment at the top of get_user.py
sort of a lie, but for now we're safe. If it becomes necessary, I
believe a more creative/hacky solution can be implemented by making
modifications to cache_loader/__init__.py.