HACKY TeacherQuizModule for MIT Splash 2010.
This in fact does not work for any other programs.

Later I want to at least specify questions from the DB,
like the surveys do--perhaps even using the same models.

No idea how to check answers cleanly, though.
Answer types might for now have to be added one by one,
as (proxy) subclasses of Answer that provide custom checking methods.

Or you could store in the DB, for each question, a chunk of python
that checks answers. Not that I think it's a good idea or anything.