Add onsite status view with simple class changes.
The goal is to get everything on one screen, and I invite comments.
Each box next to a section allows entering a student ID and hitting
Enter to submit.  If the student is already in the section, they get
taken out.  If not, they get moved from whatever classes conflict with
that section.

Sections are color coded as follows:
  Hue: Expected size of class (green=empty, red=full)
  Lightness: Percentage of check-ins (light=none, dark=all)

I'm not sure that the caching is implemented correctly but the numbers
seem to update...