Catalog preview: use 404, not 500, if not found.
If the class is not found, one of the cache key funcs dies painfully
(because it can't extract an id from the None object).

We *could* make the catalog preview fail gracefully in such a case,
but it's easier and more consistent just to raise a 404 and let
the 404-handling machinery deal with it.