Add alternative database backend for prepared statements...
To use it, change DATABASE_ENGINE to 'esp.db.prepared'.
You also need to add:
  SOUTH_DATABASE_ADAPTERS = {'default': 'south.db.postgresql_psycopg2'}

This includes a hack for connection pooling (1 connection
per process).  Some sort of persistent connections in Python
(i.e. not pgpool) are necessary because the prepared statements
cache has to be tied to the connection.  If you create a new
connection for each request as Django does by default, you end
up re-preparing all of your queries.  I hope that this part
will be improved.

In any case, performance improvements appear to be 5-10%.