Kill another sad_face cache-dumping error.
May make lottery student reg submission a bit slower as a side effect.
Given how much other computation that does, I doubt it's noticeable.

As usual, some chain of imports led to us trying to run a database
query before the caches were ready. Here's the import chain:

* esp.program.modules.models [in INSTALLED_APPS]
 > esp.program.modules.handlers
  > esp.program.modules.handlers.lotterystudentregmodule
   > esp.program.views [lsr_submit queries for Program]
    > esp.program.forms [ProgramCreationForm queries for DataTree]

When we reached the relevant queries, Django took control and imported
the rest of our apps, so that it'd be aware of all relationships in
the database.* The last app, cache_loader, also tried to import all of
our apps, so that it could be aware of all caches.

But the import of esp.program.modules.handlers had been interrupted;
so cache_loader, not being able to tell that some program modules
remained unloaded, went ahead and marked the caches as finalized.

Once this had happened, python resumed importing the rest of
esp.program.modules.handlers, and the next cache declaration that it
encountered was very sad indeed that it had been left behind.

[Footnotes]
* I think queries like (Something).objects.all() are safe since doing
  those doesn't require knowledge of reverse relationships. Even
  instantiating the objects fetched and getting the value of a foreign
  key field seems to be fine! See esp.qsd.forms.QSDMoveForm for an
  example.

See also, in case this explanation wasn't good enough:
    * Fix more sadface errors.
        (dc373eec7983ba063d1ba3f2673734d0f8bb336a)
    * Fix "Uninitialized cache." on server start.
        (d7a0a9f8b616daba25d96f15de53c8184d383a65)