Apply new license and copyright statement...
This change rewrites the legalese at the top of each
code file to reflect the license agreed upon by all of
the individual contributors.  The copyright is reassigned
to individual authors (stated in AUTHORS files, to be added
in future commits) from MIT ESP, and the license is
changed from GPL v2 to AGPL v3 to ensure full availability
of all future versions and derivatives of the code.

Also, I've converted all line endings to LF for consistency.
This was necessary since I'm on Windows and my Python script
wrote CRLF line endings when updating the licenses.