On errors, restart socket_auth.py entirely
This principally solves problems involving postgres restarting. When the
postgres connection dies, Django won't attempt to reconnect. This is
problematic for socket_auth.py, which wants to be one long-lived process.
It's not clear that Django supports telling it to reconnect, so instead
we just restart the entire process using an appropriate command-line for bash.