Correctly limit by program in get_or_create
Looking for calls to ResourceType.get_or_create, it looks like most are looking
for Classroom, Teacher Availability, or A/V (presumably wanting the "unique"
global resource with that label, and the rest probably want a resource from
the program (or a global resource if allow_global_restypes is set).

This change uses that algorithm instead of seemingly quite broken one.