Improve handling of static_resource_requests
With static resource requests (and assuming that allow_restype_creation is
unset), we need to display a form each available type --- there's no way
to add the types that we didn't start out with.

As a result, hardcoding a list of resources is kinda stupid ---
if a resource is available with the program, presumably the goal was
to make it addable. Therefore, if default_restypes is unset, add all
the available resources to the form. If default_restypes is set, we
still assume that the user knows what they're doing, strange as their
wishes may seem.

Ideally, I think that static_resource_requests should be a per-program
setting, not site-wide, and default_restypes should be ignored. But that's
a bigger change.

This commit also slightly refactors handling of resources.