Move common Javascript and CSS to elements/html; fix coteachers.
This will require most sites to edit their main.html template
and/or override elements/html.  However, I believe this change is
necessary because we want non-technical people to be able to
understand and edit templates (especially main.html).  Recently
we have had a lot of problems where important snippets of JS
or CSS were not getting loaded (breaking QSD editing, Ajax login,
or coteachers) because of missing or incorrectly ordered code
fragments somewhere in the template chain.  This commit assumes
that everyone wants to use the Dojo/YUI Javascript components
of the site and allows disabling them by overriding the appropriate
templates, rather than the other way around.