Fix migration ordering
I had earlier put in some code to delay our initial data functions
(e.g. esp.program.models.Program.install()) until a point in the
migration sequence when the appropriate tables and fields existed
in the database.  However, this code was checking all fields in the
current ORM state.  It turns out that this prevented the DataTree
from populating itself because it was missing some field in
ClassSubject that was created in a later migration.  So, I made it
possible to avoid checking all fields and merely check whether
a table exists... this is sufficient for checking whether we are ready
to run the DataTree install functions.