Add custom form inclusion support for teacher registration ...
This is a new design pattern that I wanted to try out.
BC asked for some customizations in the class creation/editing
form and I realized that chapters may want to add their own
fields (instead of having tons of fields and allowing people
to subtract them).  To avoid field proliferation in ClassSubject
and TeacherClassRegForm, I put a JSON-encoded text field on
ClassSubject and allowed additional form classes to be "stitched"
into the main form, condensing all of their data into JSON
format.  We'll see how it works out.