Restore HTML-level caching of cache_inclusion_tag
This should maintain the function signature fix of commit 58681e,
but bring back the performance improvement obtained by skipping
template rendering when the cache is hit.  There is some extra
overhead from using the second cache, but I don't see an easier
way to implement this (because the function that generates the
context dictionary is not the same as the function that generates
the HTML).