Add caching API support to template override loader
Django's cached template loader is handy since it caches compiled
templates in memory (for a bigger performance improvement over
merely caching the contents of the template).  However, there's no
way to invalidate the cache for all processes; if someone saves a
template override, only the process that handled that request
will know that the cache needs to be cleared.  The problem can
be solved by applying the caching API, albeit indirectly.

For some reason it seems to work with saving overrides, but not
deleting them.  I will see if I can expose/fix that later.