Adjust datatree test to account for unexpected URI correctness
One of our tests was checking that the DataTree's uri_correct field
was set to False after changing the name field and saving.  However,
I just recently added post_save signal sending to the DataTree.save()
function in order to fix some caching stuff.  There are a zillion
caching delete_cb functions listening on that signal, and apparently
one of them is somehow fixing the URI of the node that got saved.
I'm adjusting the test to stop complaining about that, but please
keep your ears peeled in case this is a real problem.