Add ProgramModule.inline_template field ...
This changes allows modules to have "direct links" and also display
information/forms inline with the core module page.  For example,
the TeacherClassRegModule can make a link directly to the makeaclass
page while also including the list of classes and editing options
on the main teacherreg page.  No changes are necessary to maintain
the existing behavior, but you can add an inline_template value
to an existing module.

This is only enabled for teacher modules now, but it will be trivial
to enable it for other module types in the future.