Updates to the student registration interface, for its conversion to work with HSSP registration.
In order to not accidentally break Splash registration, I made two copies of all sections of code that have changed (one is the original, one are my changes), and seperated them with appropriate if-else statements. Once the code is known to be stable, this can be removed.

Also, this commit only affects the visual behavior of lottery student reg. The code that saves the information in the database still needs to be updated.