Third version of test function for finding relationships between models.
This function takes a model and an arbitrarily long list of conditions, and returns a filter of the model based on those conditions.

Uses the first version to construct the relationships between the model and each of the conditions.

When the list is a list of instances of models, this function reduces to the second version.

This version adds support for filtering on fields of models; this is accomplished by passing a (model, field, value) 3-tuple as a condition.