Fourth version of test function for finding relationships between models.
This function takes a model and an arbitrarily long list of conditions, and returns a filter of the model based on those conditions.

Uses the first version to construct the relationships between the model and each of the conditions.

When the list is a list of instances of
 * models
 * 3-tuples
this function reduces to the third version.

This version adds support for filtering based on Django query terms; this is accomplished by passing a (model, field, query_term, value) 4-tuple as a condition.

This commit also adds a bugfix for when a user tries to call path_v2(model, instance_of_that_model) and other similar calls.