Add ConsistencyChecker for program resources/schedules...
Ted Hwa contributed a script that checks for conflicts with a particular
program.  I broke it into some functions in a new "controller" class,
esp.program.controllers.consistency.ConsistencyChecker, and made
the "Manage Class" page show the output of those functions for each
class.  Better uses should be added in the future.