Avoid setting ClassSection max_class_capacity field if possible...
This is a minor change to the behavior of the Manage Class form.
If you change the size of a class using that form, the sizes of sections
belonging to that class are automatically updated; you don't need to set
the max_class_capacity field unless you want it to be something different.