Fix (I think?) to the PersistentQueryFilter model, which was causing student schedules to miss some students.
Generating student schedules was consistently generating less students than generating nametags was.

Investigation revealed that this was because an additional filter, checking if the user is active, was being applied in the get_Q method of the PeristentQueryFilter. I changed this (for studentschedules only, as to not break anything else).