Clean up custom forms onModify() view ...
This is a minor change to reuse code a little better.
The algorithm hasn't been changed except to use updateField() on
all existing fields (ensuring attribute changes like whether
the field is required).  This will probably need sanity checking;
for example, making a field required that was non-required
might cause issues if some of the columns are empty.  South complains
about this and maybe we should just pick a policy for dealing
with it.