Add test runner for tests against existing databases....
This is a huge hack, and generally not good practice, but I'm
going to play around with it a bit.  Because our sites' exact
behavior depends on some of the data (Tags, template overrides)
in the database, it will be helpful to find any bugs caused by
interaction of the code and data in cases where the site has
been customized.