Update ProgramModule properties for the last time...
The production databases have screwed up module properties,
especially aux_calls, many of which were manually edited because
of the lack of proper data migrations (which would have been
annoying to do correctly anyway).  This code merges the
aux_calls in the database with any that are provided in the
module_properties dictionaries.

In any case, I've had enough of this and will be updating
things to introspect code instead of storing view names
in the database, when I get the chance.