Fix obscure bug with RegProfile/StudentInfo.
When saving a new RegProfile, the old StudentInfo
is fetched and updated. Except for the user.
So if, for some reason, the user was not set
in the old StudentInfo, errors would occur later.
In this case, the mailman add_list_member function
was failing, with a None user.

This should fix the bug, by setting StudentInfo.user
if it isn't already set.