Add initial support for compound linked fields...
Vishal had worked out how to collect multiple tidbits of information
using a single form field (based on Django's MultiValueField), which
was really helpful for streamlining and customizing the UI for
groups of related fields.  This is my attempt to make that functionality
work also for linked fields (e.g. if the data is in the original
schema).