Added function for finding MTO forward paths.
Not in use yet, but this could be useful, in some
cases, for selecting 'obvious' paths. For example,
StudentRegistration -> Section -> Parent Class,
which only follows ForeignKey relationships,
may very well be the only useful forward path
from StudentRegistration to ClassSubject.

But it is unclear if this method will ever be used.