Add correctness options for custom forms...
This is motivated by the desire to use custom forms for the teacher quiz.
In this case, you want some (or all) of the form fields to have 'correct'
answers that must match the values entered into the form before it can be
submitted.  Saving this data becomes useless, but oh well.

It's currently implemented on a per-field-type basis.  Drop down boxes,
radio button groups, checkboxes and short text fields are supported (this
reflects the functionality required to port existing teacher quizzes.)