Introduce the ignore_parent_name Tag.
If ignore_parent_name is set to True for a Program p,
p.niceName only returns the name of the program itself,
not the name of the parent program.

For example, if p is an instance of Delve, called Spicy Delve,
and ignore_parent_name is set to True, then the name of the
program appears as Spicy Delve, not Delve Spicy Delve.