Initial commit of mostly client-side onsite interface
The idea is to fetch a lot of information about the state of
enrollment and check-in via Ajax requests, to show the status
of all classes in a compact fashion and allow extremely fast
class changes.  This commit provides some helpful JSON views
for the backend queries as well as the template and Javascript
framework for the interface.  More to come.

This should eventually replace /onsite/[program]/[instance]/status
and maybe some other things too.