Default grade in profile creator is now null.
There was a bug where, if you set the
student_grade_options, the default grade would
be the smallest grade option. Many users missed
this box, and accidentally chose the wrong grade
(which required manual fixing in the admin panel).
Fixed the bug by supplying an initial null value.

This resolves issue #9
<https://github.com/learning-unlimited/ESP-Website/issues/9>.