Make ClassSection.num_students() respect _count_students
ClassSection has a prefetch_catalog_data() method that pre-calculates the number of students enrolled in each class, as well as the meeting times of each class, in a small number of SQL queries.  If num_students() is called with its default arguments, and if _count_students is set, it is correct (and much faster) for num_students() to simply return _count_students.