Make Program.getShirtInfo() do one query, rather than O(#teachers) queries
We now have a nice boolean field on RegistrationProfile that lets us filter by the latest profile.  Django also has shiny aggregation functions.  So we don't have to iterate through the list of users, get the profile for each user, get the teacher info off of that profile, get the T-shirt size and type off of that teacher profile, and increment a tally for that size and type.  Instead, we can just get Django to give us a table of the form

Shirt Type | Shirt Size | Count

and then perform some gymnastics to turn this into a grid with size on one axis and type on the other.

TODO:  These gymnastics are uglier than they need to be...  If we wanted to be really fancy, we could have Django return sorted output and just have the template iterate over it or something.