~20x performance improvement in rendering the cached catalog.
There's a 'for' loop in catalog.html that iterates over all of the classes, printing a short section header if at the start of a new section and then calling out to a Python function to render each actual class entry.  The Python function is fully cached.

Rather than have the 'for' loop in the template, this code instead calls the render function directly for each class, and inserts the resulting strings into an array.  Section headers are also inserted into the array at the proper locations.

Then I call ''.join(array) and ship the whole giant string as a variable to be inserted into the catalog.

Turns out ''.join is a whole lot more efficient than Django Template 'for' loops for large output.

This code also starts to add some arguments to the render functions that let people pass in various useful parameters rather than having the render function do a DB query per class to look them up.