Made progress on testing of QSD caching
I have a test that theoretically should work. Currently it's
failing because Django's caching is getting in the way. I posted
a comment to Github, waiting for a response.

I also moved some of the common selenium testing functions out of
the specific CSRF selenium test class and into the general module.
The CSRF test now just includes and uses the functions (and still
works). My new test for QSD caching uses these as well.