Force-disable setting a new CSRF cookie.
Some views, for now QSD and the catalog, are cached at the HTTP level.  That means that it's not safe for those pages to set a new CSRF cookie for users who visit them:  This will either defeat the caching or cause everyone who visits the page to be given a copy of the CSRF token of the first user who went there.

So, add a decorator for a view that will cause the csrf middleware to not set a csrf cookie for any response from that view.  And add the decorator to the catalog and to all QSD pages.

Note that this breaks non-public QSD pages:  Once one person has seen the page, anyone who knows the URL can go and see the page.  Does anyone use QSD pages that are not public, though?  I feel like we have no current uses of it, and periodic issues from people who assume that all pages are always publicly visible but who can't see them because the relevant UserBit isn't set.