Scheduling: Show correct class in conflict msg.
Say which class occupies the room, specifically at the time requested.
This should fix the messages about conflicts with classes that aren't
even from the same program.

The old behavior would say the room was occupied by whichever class was
the first to occupy a room with that name (or none at all, if the first
(room, time) pair with that name is empty). That might've been the
reason for the paranoid comment on the next line after.